requirements.txt
flask
flask-SQLAlchemy
jinja2
python-dotenv
flask-restful